bson: Move to own pkg
